pyproject.toml
[project]
name = "aicare"
version = "0.1.0"
description = "AI care — agentic Telegram bot for SG sandwich-generation eldercare coordination"
requires-python = ">=3.11"
dependencies = [
    # Core runtime
    "python-telegram-bot[ext]>=21.0",
    "apscheduler>=3.10,<4",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "supabase>=2.4",
    "openai>=1.40",
    "elevenlabs>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",

    # Config / utils
    "pydantic>=2.5",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "httpx>=0.27",

    # Domain
    "icalendar>=5.0",
    "python-dateutil>=2.9",
    "reportlab>=4.0",
    "qrcode[pil]>=7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
